Move Vanilla* classes to their respective packages, if present, don't render enchantment glint, and other clean-ups
